Added macroParam support to the time-serial-view. MacroParam can be added as main parameter or side-parameter. MacroParams can be added only by right-click popup menus at this point.

+++ libs/fmi/bizcode/generaldoc/include/NFmiEditMapGeneralDataDoc.h
@@ -493,7 +493,7 @@ public:
 	void DoAutoSaveData(void);
 	NFmiMetEditorOptionsData& MetEditorOptionsData(void);
 	// theModificationFactorCurvePoints koko tulee annetusta timebagista
-	bool DoTimeSeriesValuesModifying(boost::shared_ptr<NFmiDrawParam> &theModifiedDrawParam, int theUsedMask, NFmiTimeDescriptor& theTimeDescriptor, std::vector<double> &theModificationFactorCurvePoints, NFmiMetEditorTypes::FmiUsedSmartMetTool theEditorTool, bool fUseSetForDiscreteData, int theUnchangedValue = -1);
+	bool DoTimeSeriesValuesModifying(boost::shared_ptr<NFmiDrawParam> &theModifiedDrawParam, int theUsedMask, NFmiTimeDescriptor& theTimeDescriptor, std::vector<float> &theModificationFactorCurvePoints, NFmiMetEditorTypes::FmiUsedSmartMetTool theEditorTool, bool fUseSetForDiscreteData, int theUnchangedValue = -1);
 	bool DoCombineModelAndKlapse(void);
 	CtrlViewUtils::FmiEditorModifyToolMode ModifyToolMode(void);
 	void ModifyToolMode(CtrlViewUtils::FmiEditorModifyToolMode newState);
